+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 1.1.0
+
+Feature: Preload model when opening assistant.  
+Use *llama3.2* as default model.
+
 ## Version 1.0.2
 
 Fix: Don’t use timeout to reset continueReading toggle.